example/requirements.txt
pyzmq==14.7.0
thrift==0.9.2
gevent==1.0.1
greenlet==0.4.7
setproctitle==1.1.8
colorama==0.3.3
